Edited install/uninstall scripts, icons, and desktop file.
Corrected bug in install_rtxi.sh that pointed to the wrong directory for rtxi.desktop, and made it so that root doesn't do the pointing.
The uninstall script will delete /etc/rtxi.conf
The icon is a PNG now with a white background. Deleted the SVG file.
Edited the rtxi.desktop file so that RTXI falls under "Utilities," not "Other."