R scripts no longer need root permissions to run.
Created an .Renviron file in the home directory of the person that
 runs the script, which tells R that it is to install in a local
directory without any need for root permissions. This has one
disadvantage in that every user will have to install R packages
the first time they run this script.

This will all be unnecessary after the new live CD is made. It
will have all the R packages on it. The install checks in the R
scripts should remain, though, for people who install RTXI without
 the live CD.

An alternative to using R for the latency histogram is the gnuplot
 script that comes packaged with Xenomai.