The Channel tab of the oscilloscope will resize more prettily.
The gridLayout used to populate the Channel tab is used differently. The
widgets that went into it are now wrapped into QHBoxLayout items and then
added to the grid. This allows them to be resized as a unit and simplifies the
process of aligning all the items in the grid.

Also, spacer items have been added to the display so that when the oscilloscope
 expands, some space appears in the middle. The controls to the left are left-
aligned, and those to the right are right-aligned.

Overall, the display resizes better.