Oscilloscope is no ~10% smaller in width.
Cleanup is still needed to remove extraneous layouts and improve variable
names.

Also, I added back the line:
	QObject::connect(tabWidget,SIGNAL(currentChanged(int)),this,
	                 SLOT(showTab(int)));

It shouldn't have been removed (by me). It was probably lost in a merge commit
somewhere. Without the line, the Display tab will not show the correct
information, and the trigger combobox isn't populated with the appropriate
signals.