Cleanup of osciloscope controls done.
Widgets are no longer directly inserted into the grid layout on wither tab.
Instead, they are added to a QHBoxLayout, with a QSpacerItem placed in the
middle. There is one QHBoxLayout for each row. Once all the widgets are added,
the layouts are added to the grid layout. Additional flags were added to
widgets as needed to control the way they resize.

Resizing is now 'cleaner.'