Updating script to the new LSB compliant
The header is part of Debian policy change to move to new Init.d Script LSB (Linux Standard Base) compliant 
for more info:
https://wiki.debian.org/LSBInitScripts
The Debian policy changes are slow to introduce even for things (for an example, see #291148) which most maintainers agree are good since we have to wait first for many packages to do things in the same way before turning it into policy. Since we want to be LSB compliant, init.d scripts can be adjusted now to be LSB compliant.