Problem: RT Benchmarks needs to be compiled with runningstat, but -lrtmath isn't installed yet.
Attempt 1 (failed): Manually add the path (-L$(top_srcdir)/libs/math) to
the makefile LDFLAGS. The module will compile, but when running RTXI,
the libfile won't be able to find the lib for runningstat.

Solution: Just recompile runningstat along with the module and add a
hard link in the makefile to the directory where runningstat is.