Use xenial system libraries for qwt and hdf5:
1. Downgrading to qwt 6.1.2
2. Change -lqwt to -lqwt-qt5
3. Change linked libraries for hdf to -lhdf5 -lhdf5_hl because the
default packaging of ithe libraries is weird. You have to add the
-L/usr/lib/x86_64-linux-gnu/hdf5/serial path and then add the hdf
libraries.
4. Disable the hdf checks in configure.ac because hdf headers are not
installed in /usr/include, so the check will fail until updated to look
in the right directory.

Because a qwt package compiled with qt5 is not available in trusty,
we have to be careful about implementing checks so that qwt is compiled
in trusty but the package is use in xenial.

May have a similar issue with libhdf5 in trusty and xenial, but I
haven't checked yet.