Install libqwt-qt5-dev from repos if available. Compile if not.
Makes install_dependencies compatible with both trusty and xenial. One
thing to watch is that libqwt is renamed to libqwt-qt5. This keeps the
lib name consistent in both trusty and xenial and also adds protection
against the compiled libqwt from being rewritten if a user installs
libqwt (compiled with qt4) from the repos.