Modify the qwt source so that it installs in reasonable places.
Much of the difficulty we've had in dealing with qwt includes and
libs comes from the maintainer's decision to install everything in
/usr/local/qwt-6.x.x.

I changed the qwt code and then re-tar'ed it, so I'll write the edits
here:

1. Change the library name from qwt to qwt-qt$${QT_MAJOR_VERSION} in
src/src.pro, designer/designer.pro, and qwt.prf. This changes the name
of the library from libqwt to libqwt-qt5.

2. Change the install directories in qwtconfig.pri. For example, I
changed:

  QWT_INSTALL_PREFIX  = /usr/local
  QWT_INSTALL_HEADERS = $${QWT_INSTALL_PREFIX}/include/qwt

I also changed the Makefiles to be consistent with the new paths.